Some warnings clang wouldn't shut up about.

+++ ONScripterLabel_effect_cascade.cpp
@@ -40,8 +40,8 @@ void ONScripterLabel::effectCascade( char *params, int duration )
     };
 
     SDL_Surface *src_surface, *dst_surface;
-    SDL_Rect src_rect={0, 0, screen_width, screen_height};
-    SDL_Rect dst_rect={0, 0, screen_width, screen_height};
+    SDL_Rect src_rect={0, 0, (Uint16)screen_width, (Uint16)screen_height};
+    SDL_Rect dst_rect={0, 0, (Uint16)screen_width, (Uint16)screen_height};
     int mode, width, start, end;
  
     if (params[0] == 'u')
